+++ crucible-llvm-syntax/CHANGELOG.md
@@ -1,6 +1,6 @@
 # next
 
-* Add overrides for string-manipulation `{read,write}-bytes` and
+* Add overrides for string-manipulation: `{read,write}-bytes` and
   `{read,write}-c-string`.
 
 # 0.1 -- 2024-02-05